Fixed compiling error.

+++ include/pico_constants.h
@@ -8,7 +8,7 @@ See LICENSE and COPYING for usage.
 /* Included from pico_config.h */
 /** Endian-dependant constants **/
 
-extern volatile unsigned long long pico_tick;
+extern volatile uint64_t pico_tick;
 
 #ifdef PICO_BIGENDIAN
